🔨 Refactor: 겹치는 코드 컴포넌트로 분리
