+++ src/pages/addboard.tsx
@@ -4,14 +4,15 @@ import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 import Link from 'next/link';
 import { AxiosError } from 'axios';
+import { StyledToastContainer } from '@/styles/ToastStyle';
 import CommonButton from '@/components/common/CommonButton';
 import dateToString from '@/utils/dateToString';
-import { extractFirstImgSrc, roundAttributes } from '@/utils/quillHtmlHandler';
+import { refineHTMLContent } from '@/utils/quillHtmlHandler';
 import { type ArticleFormData } from '@/types/apiType';
 import { postArticle } from '@/lib/apis/article/articleApi.api';
 import ToastSelect from '@/components/common/ToastSelect';
 import { OTHER_TYPE_ERROR_TEXT } from '@/constants/otherTypeErrorText';
-import { StyledToastContainer } from '@/styles/ToastStyle';
+import BoardInfoForm from '@/components/AddBoard/BoardInfoForm';
 import 'react-toastify/dist/ReactToastify.css';
 
 const ReactQuillWrapper = dynamic(import('@/components/AddBoard/QuillEditor'), {
@@ -21,29 +22,19 @@ const ReactQuillWrapper = dynamic(import('@/components/AddBoard/QuillEditor'), {
   },
 });
 
-const TITLE_MAX_LEN = 30;
-
 const AddBoard = () => {
   const [title, setTitle] = useState('');
   const [content, setContent] = useState('');
   const [contentLength, setContentLength] = useState({ withSpaces: 0, withoutSpaces: 0 });
   const [isValid, setIsValid] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
+
   const router = useRouter();
 
   const isButtonActive = isValid && !isLoading;
   const isButtonDisabled = !isValid || isLoading;
 
-  const refineHTMLContent = (initContext: string) => {
-    let newContent = '';
-
-    const firstImageSrc = extractFirstImgSrc(initContext);
-    newContent = initContext.replace(/cursor: (nesw|nwse)-resize;/g, '');
-    newContent = roundAttributes(newContent);
-
-    return { firstImageSrc, newContent };
-  };
-
+  // 게시물 작성하기
   const handleSubmit = async () => {
     setIsLoading(true);
     const { firstImageSrc, newContent } = refineHTMLContent(content);
@@ -61,7 +52,7 @@ const AddBoard = () => {
       const response = await postArticle(boardData);
       ToastSelect({
         type: 'check',
-        message: '게시물이 수정되었습니다',
+        message: '게시물이 작성되었습니다',
         onClose: () => {
           router.push(`/board/${response.id}`);
         },
@@ -76,6 +67,7 @@ const AddBoard = () => {
     }
   };
 
+  // 게시물 내용 입력
   const handleInputContent = (
     content: string,
     length: { withSpaces: number; withoutSpaces: number },
@@ -84,6 +76,7 @@ const AddBoard = () => {
     setContentLength(length);
   };
 
+  // 게시물 유효성 검사
   useEffect(() => {
     setIsValid(title.trim().length > 0 && content.trim().length > 0);
   }, [title, content]);
@@ -107,29 +100,7 @@ const AddBoard = () => {
         <span className="text-xs-regular text-gray-400 md:text-lg-regular">
           등록일 {dateToString(new Date())}
         </span>
-        <div>
-          <div className="mt-1 border-t" />
-          <div className="flex items-center justify-between gap-2">
-            <input
-              onChange={(e) => {
-                setTitle(e.target.value);
-              }}
-              value={title}
-              className="flex-1 rounded-sm py-3 text-lg-medium outline-none md:text-xl-medium"
-              placeholder="제목을 입력해주세요"
-              maxLength={TITLE_MAX_LEN}
-            />
-            <span className="text-xs-medium md:text-md-medium">
-              {title.length}/<span className="text-primary-green-200">{TITLE_MAX_LEN}</span>
-            </span>
-          </div>
-          <div className="border-t" />
-        </div>
-        <span className="text-md-medium md:text-lg-medium">
-          공백포함 : 총<span className="text-primary-green-200"> {contentLength.withSpaces}</span>자
-          | 공백제외: 총
-          <span className="text-primary-green-200"> {contentLength.withoutSpaces}</span>자
-        </span>
+        <BoardInfoForm title={title} setTitle={setTitle} contentLength={contentLength} />
         <ReactQuillWrapper setContent={handleInputContent} content={content} />
       </main>
       <Link href="/boards" rel="preload">
