+++ src/pages/board/[boardId]/edit.tsx
@@ -6,8 +6,9 @@ import Link from 'next/link';
 import { AxiosError } from 'axios';
 import CommonButton from '@/components/common/CommonButton';
 import dateToString from '@/utils/dateToString';
-import { extractFirstImgSrc, roundAttributes } from '@/utils/quillHtmlHandler';
-import { type ArticleFormData } from '@/types/apiType';
+import BoardInfoForm from '@/components/AddBoard/BoardInfoForm';
+import { refineHTMLContent } from '@/utils/quillHtmlHandler';
+import { type DateType, type ArticleFormData } from '@/types/apiType';
 import { changeDetailArticle, getDetailArticle } from '@/lib/apis/article/articleApi.api';
 import ToastSelect from '@/components/common/ToastSelect';
 import { OTHER_TYPE_ERROR_TEXT } from '@/constants/otherTypeErrorText';
@@ -32,12 +33,11 @@ const StyledToastContainer = dynamic(
   },
 );
 
-const TITLE_MAX_LEN = 30;
-
 const EditBoard = () => {
   const [title, setTitle] = useState('');
   const [content, setContent] = useState('');
   const [contentLength, setContentLength] = useState({ withSpaces: 0, withoutSpaces: 0 });
+  const [createAt, setCreateAt] = useState<DateType>(new Date());
   const [isValid, setIsValid] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
 
@@ -47,16 +47,7 @@ const EditBoard = () => {
   const isButtonActive = isValid && !isLoading;
   const isButtonDisabled = !isValid || isLoading;
 
-  const refineHTMLContent = (initContext: string) => {
-    let newContent = '';
-
-    const firstImageSrc = extractFirstImgSrc(initContext);
-    newContent = initContext.replace(/cursor: (nesw|nwse)-resize;/g, '');
-    newContent = roundAttributes(newContent);
-
-    return { firstImageSrc, newContent };
-  };
-
+  // 게시물 수정하기
   const handleSubmit = async () => {
     setIsLoading(true);
     const { firstImageSrc, newContent } = refineHTMLContent(content);
@@ -64,11 +55,8 @@ const EditBoard = () => {
     const boardData: ArticleFormData = {
       title,
       content: newContent,
+      image: firstImageSrc ? firstImageSrc : undefined,
     };
-
-    if (firstImageSrc) {
-      boardData.image = firstImageSrc;
-    }
 
     try {
       const response = await changeDetailArticle(Number(boardId), boardData);
@@ -89,6 +77,7 @@ const EditBoard = () => {
     }
   };
 
+  // 게시물 내용 입력
   const handleInputContent = (
     content: string,
     length: { withSpaces: number; withoutSpaces: number },
@@ -97,29 +86,32 @@ const EditBoard = () => {
     setContentLength(length);
   };
 
+  // 게시물 유효성 검사
   useEffect(() => {
     setIsValid(title.trim().length > 0 && content.trim().length > 0);
   }, [title, content]);
 
+  // 수정할 게시물 가져오기
   useEffect(() => {
     const fetchBoardData = async () => {
-      if (boardId) {
-        try {
-          const response = await getDetailArticle(Number(boardId));
-          setTitle(response.title);
-          setContent(response.content);
-        } catch (error) {
-          if (error instanceof AxiosError) {
-            if (error.status === 404) {
-              router.push('/404');
-            } else {
-              router.push('/500');
-            }
+      try {
+        const response = await getDetailArticle(Number(boardId));
+        setTitle(response.title);
+        setContent(response.content);
+        setCreateAt(response.createdAt);
+      } catch (error) {
+        if (error instanceof AxiosError) {
+          if (error.status === 404) {
+            router.push('/404');
+          } else {
+            router.push('/500');
           }
         }
       }
     };
-    fetchBoardData();
+    if (boardId) {
+      fetchBoardData();
+    }
   }, [boardId, router]);
 
   return (
@@ -140,31 +132,9 @@ const EditBoard = () => {
           </CommonButton>
         </div>
         <span className="text-xs-regular text-gray-400 md:text-lg-regular">
-          등록일 {dateToString(new Date())}
+          등록일 {dateToString(createAt)}
         </span>
-        <div>
-          <div className="mt-1 border-t" />
-          <div className="flex items-center justify-between gap-2">
-            <input
-              onChange={(e) => {
-                setTitle(e.target.value);
-              }}
-              value={title}
-              className="flex-1 rounded-sm py-3 text-lg-medium outline-none md:text-xl-medium"
-              placeholder="제목을 입력해주세요"
-              maxLength={TITLE_MAX_LEN}
-            />
-            <span className="text-xs-medium md:text-md-medium">
-              {title.length}/<span className="text-primary-green-200">{TITLE_MAX_LEN}</span>
-            </span>
-          </div>
-          <div className="border-t" />
-        </div>
-        <span className="text-md-medium md:text-lg-medium">
-          공백포함 : 총<span className="text-primary-green-200"> {contentLength.withSpaces}</span>자
-          | 공백제외: 총
-          <span className="text-primary-green-200"> {contentLength.withoutSpaces}</span>자
-        </span>
+        <BoardInfoForm title={title} setTitle={setTitle} contentLength={contentLength} />
         <ReactQuillWrapper setContent={handleInputContent} content={content} />
       </main>
       <Link href={`/board/${boardId}`} rel="preload">
